% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mengbo_colors.R
\docType{data}
\name{mengbo_colors}
\alias{mengbo_colors}
\title{Basic colors}
\format{
An object of class \code{character} of length 29.
}
\usage{
mengbo_colors
}
\description{
Basic colors
}
\examples{
mengbo_colors
}
\keyword{datasets}
