% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mengbo_cols.R
\name{mengbo_cols}
\alias{mengbo_cols}
\title{Function to extract mengbo colors as hex codes}
\usage{
mengbo_cols(...)
}
\arguments{
\item{...}{Character names of mengbo_colors}
}
\description{
Function to extract mengbo colors as hex codes
}
\examples{
mengbo_cols()
mengbo_cols("red")
mengbo_cols("red", "blue")
}
