% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mengbo_pal.R
\name{mengbo_pal}
\alias{mengbo_pal}
\title{Return function to interpolate a mengbo color palette}
\usage{
mengbo_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in drsimonj_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\value{
Another function.
}
\description{
This function gets a pallete by name from the list ("main" by default),
has a boolean condition determining whether to reverse the order or not, and
additional arguments to pass on to colorRampPallete()
(such as an alpha value).
}
\note{
Changes or new color palettes are added in this list.
}
\examples{
mengbo_pal("main", alpha = 0.8)
mengbo_pal("cool")(10)
}
