% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mengbo_palettes.R
\docType{data}
\name{mengbo_palettes}
\alias{mengbo_palettes}
\title{Basic palettes}
\format{
An object of class \code{list} of length 12.
}
\usage{
mengbo_palettes
}
\description{
Basic palettes
}
\examples{
mengbo_palettes$main
}
\keyword{datasets}
