% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_mengbo.R
\name{scale_color_mengbo}
\alias{scale_color_mengbo}
\title{Color scale constructor for mengbo colors}
\usage{
scale_color_mengbo(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in drsimonj_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is}
}
\description{
Color scale constructor for mengbo colors
}
\examples{
# Color by discrete variable using default palette
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
geom_point(size = 4) + scale_color_mengbo()
# Color by numeric variable with cool palette
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Sepal.Length)) +
geom_point(size = 4) + scale_color_mengbo(discrete = FALSE, palette = "cool")

}
