% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_mengbo.R
\name{scale_fill_mengbo}
\alias{scale_fill_mengbo}
\title{Fill scale constructor for mengbo colors}
\usage{
scale_fill_mengbo(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in mengbo_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is
TRUE or FALSE}
}
\description{
Fill scale constructor for mengbo colors
}
\examples{
# Fill by discrete variable with different palette + remove legend (guide)
ggplot(mpg, aes(manufacturer, fill = manufacturer)) + geom_bar() +
theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
scale_fill_mengbo(palette = "mixed", guide = "none")

}
