% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mengbo.R
\name{theme_mengbo}
\alias{theme_mengbo}
\title{Plotting theme modified from theme_bw}
\usage{
theme_mengbo(
  base_size = 14,
  base_family = "",
  base_line_size = base_size/20,
  base_rect_size = base_size/20,
  clean_background = TRUE
)
}
\arguments{
\item{base_size}{base font size, given in pts}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}

\item{clean_background}{Logical, whether to keep the grid guidelines}
}
\description{
Plotting theme modified from theme_bw
}
\details{
\describe{
\item{`theme_mengbo()`}{
Mengbo's own ggplot theme modified from the classic dark-on-light ggplot2
theme.
}}
}
\examples{
mtcars2 <- within(mtcars, {
  vs <- factor(vs, labels = c("V-shaped", "Straight"))
  am <- factor(am, labels = c("Automatic", "Manual"))
  cyl  <- factor(cyl)
  gear <- factor(gear)
})

p1 <- ggplot(mtcars2) +
  geom_point(aes(x = wt, y = mpg, colour = gear)) +
  labs(
    title = "Fuel economy declines as weight increases",
    subtitle = "(1973-74)",
    caption = "Data from the 1974 Motor Trend US magazine.",
    # tag = "Figure 1",
    x = "Weight (1000 lbs)",
    y = "Fuel economy (mpg)",
    colour = "Gears"
  )

p1 + theme_mengbo()

}
