% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wehi_colors.R
\docType{data}
\name{wehi_colors}
\alias{wehi_colors}
\title{WEHI colors}
\format{
An object of class \code{character} of length 9.
}
\usage{
wehi_colors
}
\description{
WEHI colors
}
\examples{
wehi_colors
}
\keyword{datasets}
